requirements.txt
django==4.1
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1